refactor: Change issue key
