+++ processor/src/test/java/org/mapstruct/ap/test/bugs/_3884/Issue3884Mapper.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.bugs._3865;
+package org.mapstruct.ap.test.bugs._3884;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
@@ -14,8 +14,8 @@ import org.mapstruct.factory.Mappers;
  * Mapper for testing null value property mapping strategy with Map properties.
  */
 @Mapper(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.SET_TO_DEFAULT)
-public interface Issue3865Mapper {
-    Issue3865Mapper INSTANCE = Mappers.getMapper( Issue3865Mapper.class );
+public interface Issue3884Mapper {
+    Issue3884Mapper INSTANCE = Mappers.getMapper( Issue3884Mapper.class );
 
     void update(@MappingTarget DestinationType destination, SourceType source);
 }
