+++ processor/src/test/java/org/mapstruct/ap/test/bugs/_3884/SourceType.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.bugs._3865;
+package org.mapstruct.ap.test.bugs._3884;
 
 import java.util.ArrayList;
 import java.util.HashMap;
